Varying Results Fixed
Assumed Fix:

By repeatedly using the same request variable, some ISEND/RECV pairs were being skipped over or accidentally skipped. As there are 2-3 uses of this, it makes sense that the results usually had 2-3 variants. By substituting all of these for MPI_SENDRECV (except one, which also makes sense as now there is only one instance of it), which just does the job of MPI ISEND and RECV, the problem is resolved. Next step is to make any nprocs work