Problem Resolved1
Issue one: Cut off was too big for some processor numbers. This has been resolved

Issue Two: In the MPI_SENDRECV, i hadn't swapped around the rank source and locations for the 2nd set of paritcles per loop meaning ti was sending before and after particles in the same direction. This is why it worked for 1,2,4,8 processors where there are at most 2 processes on each